+++ OOP-ShapesProject/Rect.cs
@@ -6,7 +6,7 @@ namespace OOP_ShapesProject {
     class Rect: Quad {
 
         public int Area() {
-            return this.side1 * this.side2;
+            return (int)(this.side1 * this.side2);
         }
 
         public Rect(int side_x, int side_y)
